feat: Implement payload relaying

+++ experiments/Lavalink4NET.Experiments.Receive/Connections/Payloads/Converters/HeartbeatIntervalJsonConverter.cs
@@ -7,11 +7,11 @@ internal sealed class HeartbeatIntervalJsonConverter : JsonConverter<TimeSpan>
 {
     public override TimeSpan Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
-        return TimeSpan.FromMilliseconds(reader.GetUInt64());
+        return TimeSpan.FromMilliseconds(reader.GetDouble());
     }
 
     public override void Write(Utf8JsonWriter writer, TimeSpan value, JsonSerializerOptions options)
     {
-        writer.WriteNumberValue((ulong)value.TotalMilliseconds);
+        writer.WriteNumberValue(value.TotalMilliseconds);
     }
 }
